requirements.txt
streamlit
pdfplumber
gTTS
sentence-transformers
chromadb
openai
google-generativeai
requests
Pillow
reportlab

